Adding functionality for creating and updating jobseeker's profile

+++ hireme-web-app/src/main/java/com/hireme/model/User.java
@@ -1,3 +1,11 @@
+/**
+ * Model class for Role entity.
+ * 
+ * @author	Sushant Vairagade
+ * @version 1.0
+ * @Since	11-05-2017
+ */
+
 package com.hireme.model;
 
 import java.util.Set;
@@ -25,7 +33,7 @@ public class User {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "user_id")
-	private int id;
+	private long id;
 	
 	@Column(name = "email")
 	@Email(message = "*Please provide a valid Email")
@@ -53,11 +61,11 @@ public class User {
 	@JoinTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_id"))
 	private Set<Role> roles;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
